+++ dlt/common/libs/pandas.py
@@ -1,0 +1,7 @@
+from dlt.common.exceptions import MissingDependencyException
+
+try:
+    import pandas
+    from pandas.io.sql import _wrap_result
+except ModuleNotFoundError:
+    raise MissingDependencyException("DLT Pandas Helpers", ["pandas"])
