+++ dlt/destinations/sql_client.py
@@ -172,7 +172,7 @@ class DBApiCursorImpl(DBApiCursor):
         return [c[0] for c in self.native_cursor.description]
 
     def df(self, chunk_size: int = None, **kwargs: Any) -> Optional[DataFrame]:
-        from dlt.helpers.pandas_helper import _wrap_result
+        from dlt.common.libs.pandas import _wrap_result
 
         columns = self._get_columns()
         if chunk_size is None:
